cdk-plant-disease/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
boto3>=1.26.0
numpy>=1.22.0
tensorflow-cpu>=2.10.0
Pillow>=9.0.0
